rename gluu-client-auth to gluu-oauth-pep, #258
